Encrypt with AES Algorithm
Implemented file encryption feature in index.ejs. This feature allows users to upload a file and select an encryption algorithm to encrypt the file. Currently, the feature supports AES encryption algorithm.

The encryption process is handled by a form submission which triggers the encryptData function. This function sends a POST request to the /encrypt-collection endpoint with the file and selected encryption algorithm as the payload.

The server-side logic for this endpoint is defined in the encryptController file. Once the file is successfully encrypted, the server responds with the filename of the encrypted file. The client-side script then generates download links for the encrypted file and the key file.